feat: add mock data for labs
