+++ apps/community/src/app/api/mock/labs/route.ts
@@ -1,0 +1,5 @@
+import { labs } from './data';
+
+export function GET() {
+  return Response.json({ data: labs });
+}
