+++ apps/community/src/app/api/mock/route.ts
@@ -1,0 +1,3 @@
+export async function GET() {
+  return Response.json({ message: 'Mock API working in development' });
+}
