+++ apps/community/src/app/lab/components/lab-card.css.ts
